requirements.txt
imgaug==0.4.0
opencv-python==4.5.3.56
pandas==1.3.1
Pillow==8.4.0
Shapely==1.7.1
sklearn==0.0
tensorflow==2.6.0
google-cloud-storage>=2
